feat: Added option to search movie by cast.
Added an option that allows the user to search for a movie based on the cast members. When a match is found the movie name is displayed and the cast member is added to a list containing all others cast member matches. The cast member list is also displayed.

issue #6